+++ cx-portal/src/types/Constants.ts
@@ -79,6 +79,8 @@ export enum PAGES {
   INTRODUCTION_APP_PROVIDER = 'companyrolesappprovider',
   INTRODUCTION_SERVICE_PROVIDER = 'companyrolesserviceprovider',
   INTRODUCTION_CONFIRMITY_BODY = 'companyrolesconfirmitybody',
+  USE_CASE = 'usecase',
+  USE_CASE_TRACABILITY = 'usecasetraceablity',
 }
 
 export enum OVERLAYS {