+++ CHANGELOG.md
@@ -25,6 +25,8 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
    * View update in Details overlay
 * Notification
    * Bug fix - Pick appropriate title from locale file
+* Use case
+   * Static page content added
 * Feature - App Change Process
    * Created Deactivate Sub-Menu Component in App Overview Page
    * Created Deactivate page and add functionality